requirements.txt
asgiref==3.8.1
Django==5.0.6
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
importlib_resources==6.4.0
mysqlclient==2.2.4
PyJWT==2.8.0
pytz==2024.1
rest-framework-simplejwt==0.0.2
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
zipp==3.18.2
boto3~=1.34.108
twilio~=9.0.5
python-dotenv~=1.0.1
